requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
chromadb>=0.4.0
cryptography>=3.4.8
scikit-learn>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
